moved ErrorStateWithButton to common
