+++ frontend/src/components/common/ErrorStateWithButton.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Center } from "@chakra-ui/react";
 
-import MessageContainer from "../common/MessageContainer";
+import MessageContainer from "./MessageContainer";
 
 type ErrorStateWithButtonProps = {
   buttonText: string;
