+++ frontend/src/components/pages/teacher/ClassroomsPage.tsx
@@ -21,8 +21,8 @@ import AuthContext from "../../../contexts/AuthContext";
 import type { Classroom } from "../../../types/ClassroomTypes";
 import { TabEnumClassroom } from "../../../types/ClassroomTypes";
 import ClassroomCard from "../../classrooms/ClassroomCard";
-import ErrorStateWithButton from "../../classrooms/ErrorStateWithButton";
 import ErrorState from "../../common/ErrorState";
+import ErrorStateWithButton from "../../common/ErrorStateWithButton";
 import HeaderWithButton from "../../common/HeaderWithButton";
 import LoadingState from "../../common/LoadingState";
 import Pagination from "../../common/table/Pagination";
